Use oauth to auth with gh api
Instead of 60 requests per hour, this allows 5000. Given the amount of
libraries we need to get follower information for this is way better.

In order to use the auth you will need to define a `config.js` file.
Simply copy the provided template, generate oauth key through GitHub UI
and pass it to the utility using the configuration parameter.

In case configuration parameter is not provided it will generate dummy
data instead (handy for testing).