Improve frontend structure and update Angular
Note that ng-if seems glitchy in RC3 so I ended up replacing those with
ng-show. As a result animations were disabled.

While at it I removed a couple of redundant dependencies from the
project and added routing dependency. That is useful particularly if we
want to do something more special like linking to a specific library.